worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "y-agent-worker"
version = "0.4.0"
description = "Background tasks for y-agent"
requires-python = ">=3.10"
dependencies = [
    "y-agent-storage",
    "y-agent-agent",
    "python-dotenv",
    "loguru>=0.7.3",
    "celery>=5.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/worker"]

[tool.uv.sources]
y-agent-storage = { workspace = true }
y-agent-agent = { workspace = true }
